Move the ActionLog component
