+++ frontend/components/ActionLog.tsx
@@ -1,14 +1,14 @@
 import ActionLog from "@/classes/ActionLog"
-import SelectFactionLeaderNotification from "./ActionLog_SelectFactionLeader"
-import FaceMortalityNotification from "./ActionLog_FaceMortality"
-import TemporaryRomeConsulNotification from "./ActionLog_TemporaryRomeConsul"
-import NewTurnNotification from "./ActionLog_NewTurn"
-import NewFamilyNotification from "./ActionLog_NewFamily"
-import NewWarNotification from "./ActionLog_NewWar"
-import MatchedWarNotification from "./ActionLog_MatchedWar"
-import NewEnemyLeaderNotification from "./ActionLog_NewEnemyLeader"
-import MatchedEnemyLeaderNotification from "./ActionLog_MatchedEnemyLeader"
-import NewSecretNotification from "./ActionLog_NewSecret"
+import SelectFactionLeaderNotification from "./actionLogs/ActionLog_SelectFactionLeader"
+import FaceMortalityNotification from "./actionLogs/ActionLog_FaceMortality"
+import TemporaryRomeConsulNotification from "./actionLogs/ActionLog_TemporaryRomeConsul"
+import NewTurnNotification from "./actionLogs/ActionLog_NewTurn"
+import NewFamilyNotification from "./actionLogs/ActionLog_NewFamily"
+import NewWarNotification from "./actionLogs/ActionLog_NewWar"
+import MatchedWarNotification from "./actionLogs/ActionLog_MatchedWar"
+import NewEnemyLeaderNotification from "./actionLogs/ActionLog_NewEnemyLeader"
+import MatchedEnemyLeaderNotification from "./actionLogs/ActionLog_MatchedEnemyLeader"
+import NewSecretNotification from "./actionLogs/ActionLog_NewSecret"
 
 interface NotificationItemProps {
   notification: ActionLog
